Turn - phase check removal, new_turn removal
Turn will not check in which phase the game is. It will only check whether it is possible to buy or play a given card. It is possible to end turn at any point and do it multiple times.
Instead of new_turn a new Turn instance should be used.